refactor: amazon.tsを汎用メソッドとして移動
別phaseでも使用したいメソッドがあるため
